pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-integrity-lab"
version = "0.1.0"
description = "A real-world environment for training agents to clean and format datasets using Pandas."
authors = [
    { name = "Likith D T", email = "likith.dt77@gmail.com" }
]
dependencies = [
    "openenv-core>=0.2.2",
    "pandas",
    "pydantic>=2.0.0",
    "fastapi",
    "uvicorn"
]
readme = "README.md"
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
server = "server.app:main"

[tool.uv]
managed = true